+++ website/src/_services/charts/line-chart/functions.ts
@@ -1,8 +1,9 @@
 import * as d3 from 'd3'
 
+import { generateHorizontalLegend, getLegendGroupNames } from '..'
 import { LineChartConfig, LineChartSeries } from '.'
 
-export const generateChart = (datasets: LineChartSeries[], config: LineChartConfig): SVGSVGElement | null => {
+export const generateChart = (datasets: LineChartSeries[], config: LineChartConfig, isExported: boolean = false): SVGSVGElement | null => {
   // Prepare data
   const yBounds = [0, datasets.reduce((acc, cur) => Math.max(acc, Math.max(...Object.values(cur.data))), 0)]
   const xBounds = config.xBounds ?? getXBoundsFromDatasets(datasets)
@@ -65,6 +66,12 @@ export const generateChart = (datasets: LineChartSeries[], config: LineChartConf
       .attr('transform', x => `translate(${xScale(x)}, ${yScale(data[x] ?? 0)})`)
       .style('fill', dataset.color)
   })
+
+  if (isExported) {
+    const labels = getLegendGroupNames(datasets.length)
+    const colors = datasets.map(d => d.color)
+    generateHorizontalLegend(config.width, config.height - config.margins.bottom, labels, colors, svg)
+  }
 
   return svg.node()
 }