+++ website/src/species-richness/components/species-richness-by-time/species-richness-by-time.ts
@@ -5,7 +5,9 @@ import { Prop } from 'vue-property-decorator'
 import { TimeBucket } from '~/api/species-richness-service'
 import { svgToPngData } from '~/charts'
 import { generateChart, LineChartComponent, LineChartConfig, LineChartSeries } from '~/charts/line-chart'
+import { getExportGroupName } from '~/dataset-filters/functions'
 import { downloadPng } from '~/utils/file'
+
 
 const BUCKETS_TO_X_BOUNDS: Partial<Record<TimeBucket, [number, number]>> = {
   hour: [0, 23],
@@ -41,11 +43,12 @@ export default class SpeciesRichnessByTime extends Vue {
   }
 
   async downloadChart (): Promise<void> {
-    const exportConfig = { ...this.config, width: 800, height: 450 }
-    const svg = await generateChart(this.datasetsForSelectedBucket, exportConfig)
+    const margins = { ...this.config.margins, bottom: 80 }
+    const exportConfig = { ...this.config, margins, width: 800, height: 450 }
+    const svg = await generateChart(this.datasetsForSelectedBucket, exportConfig, true)
     if (!svg) return
 
     const png = await svgToPngData({ svg, ...exportConfig })
-    downloadPng(png, `${this.domId}-${this.selectedBucket}`) // TODO 107 - Better filename
+    downloadPng(png, getExportGroupName(`${this.domId}-${this.selectedBucket}`))
   }
 }