+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * **features:** Add export ellipsis button(#131)
 * **feature:** Research about API to get species information (#163)
 * **features:** Activity patterns page detection frequency and occupancy rate metrics (#130)
+* **features:** Add legend to bar chart and line chart export (#109)
 
 * **fixed:** Update SR line-chart to correctly display SR instead of detections (#172)